LPS-87327 Rename the property：Because the property only takes effect when connects a new database on startup. Adding sample data should be controlled by property setting, not setup wizard.
