+++ portal-impl/src/com/liferay/portal/internal/servlet/MainServlet.java
@@ -82,7 +82,7 @@ import com.liferay.portal.servlet.EncryptedServletRequest;
 import com.liferay.portal.servlet.I18nServlet;
 import com.liferay.portal.servlet.filters.absoluteredirects.AbsoluteRedirectsResponse;
 import com.liferay.portal.servlet.filters.i18n.I18nFilter;
-import com.liferay.portal.setup.SetupWizardSampleDataUtil;
+import com.liferay.portal.setup.SetupSampleDataUtil;
 import com.liferay.portal.struts.PortalRequestProcessor;
 import com.liferay.portal.struts.StrutsUtil;
 import com.liferay.portal.struts.TilesUtil;
@@ -436,11 +436,9 @@ public class MainServlet extends HttpServlet {
 			_log.error(e, e);
 		}
 
-		if (StartupHelperUtil.isDBNew() &&
-			PropsValues.SETUP_WIZARD_ADD_SAMPLE_DATA) {
-
+		if (StartupHelperUtil.isDBNew() && PropsValues.SETUP_ADD_SAMPLE_DATA) {
 			try {
-				SetupWizardSampleDataUtil.addSampleData(
+				SetupSampleDataUtil.addSampleData(
 					PortalInstances.getDefaultCompanyId());
 			}
 			catch (Exception e) {
