+++ portal-impl/src/portal.properties
@@ -5920,6 +5920,17 @@
     multi.value.map.com.liferay.portal.convert.ConvertPermissionAlgorithm.convertRoles=0
 
 ##
+## Setup Sample Data
+##
+
+    # Set this property to true to automatically add sample data to a new
+    # database on startup. 
+    #
+    # Env: LIFERAY_SETUP_PERIOD_ADD_PERIOD_SAMPLE_PERIOD_DATA
+    #
+    setup.add.sample.data=false
+
+##
 ## Setup Wizard
 ##
 
@@ -5943,15 +5954,6 @@
     setup.database.url[postgresql]=jdbc:postgresql://localhost:5432/lportal
     setup.database.url[sqlserver]=jdbc:sqlserver://localhost;databaseName=lportal
     setup.database.url[sybase]=jdbc:sybase:Tds:localhost:5000/lportal
-
-    #
-    # Set this property to true to automatically add sample data to a new
-    # database on startup. This will happen even if the Setup Wizard is
-    # disabled.
-    #
-    # Env: LIFERAY_SETUP_PERIOD_WIZARD_PERIOD_ADD_PERIOD_SAMPLE_PERIOD_DATA
-    #
-    setup.wizard.add.sample.data=false
 
     #
     # Set this property to true if the Setup Wizard should be displayed the
